sql/pgwire: Avoid uint32 -> int32 overflows which could cause panics
Fuzz testing revealed that it was possible for a uint32 overflow when a
client-provided uint32 value was cast to an int32. This could result in
a panic if this value was used as a slice index (in readBuffer.getBytes).
This change fixes this bug (but does not add a test case, as the setup
required for a proper exploit was a little tricky).

+++ sql/pgwire/encoding.go
@@ -138,11 +138,11 @@ func (b *readBuffer) getUint16() (uint16, error) {
 	return v, nil
 }
 
-func (b *readBuffer) getInt32() (int32, error) {
+func (b *readBuffer) getUint32() (uint32, error) {
 	if len(b.msg) < 4 {
 		return 0, errors.Errorf("insufficient data: %d", len(b.msg))
 	}
-	v := int32(binary.BigEndian.Uint32(b.msg[:4]))
+	v := binary.BigEndian.Uint32(b.msg[:4])
 	b.msg = b.msg[4:]
 	return v, nil
 }
